Bug 12334: Make it easy to hide parts of material type and format information in results
With the system preferences DisplayIconsXSLT and DisplayOPACiconsXSLT we control
the visibility of the material type, format and audience information in staff
and OPAC.

Sometimes a library might only want to hide a part of that information - for
example, hide audience but keep the material type icon.

This patch adds CSS classes to make it easier to style this section of the page
and hide parts of the information.

To test:
- Verify that OPAC and staff result lists for various types of materials
  still display nicely.
- Take a look at the HTML and verify, that label and text are now
  wrapped into a new span with a results_* class.
- Try hiding a part of the information, for example in OpacUserCSS:

.results_material_type {
  display none;
}

Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>
Works as described... but for a missing colon on example CSS

.results_material_type {
  display: none;
}

No koha-qa errors

Signed-off-by: Brendan Gallagher <brendan@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>