Bug 12740 - It should not be possible to store empty news
This patch marks the title field as 'required' and it is not
longer possible to store a news without having at least a title.

To test:
Got to Tools->News
Create news or edit existing news
Field "Title" should have a red label and should be marked as "required"
It should not be possible to save with empty field "Title"

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>

NOTE: Only keeps the user from doing it.
      Fails to handle the fake the url call case.
https://demo.library-admin.kohasystem.ca/cgi-bin/koha/tools/koha-news.pl?op=add&lang=&branch=&title=&from=&to=&number=&new=

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>