Bug 13042 - Move budget action links into menu
The series of links in the "actions" column of the budgets
administration summary table are not very readable as they are, all
strung together: Edit Delete Duplicate Close Add fund.

This patch moves the links into an "Actions" menu, similar to the change
made by Bug 10615 in Reports.

To test you should have multiple active and inactive budgets. Apply the
patch and go to Administration -> Budgets. Confirm that each menu item
links to the correct action. Test the menu for both active and inactive
budgets.

Signed-off-by: Paola Rossi <paola.rossi@cineca.it>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Nice improvement on template level, works as described.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>