Bug 13095: An email will be sent shortly
This patch adjusts two instances where Koha says that an email has
been sent while it is just enqueued (put in the message queue). The
crontab example still suggests to run process_message_queue once an hour
and the manual even speaks about 1-4 hours.

In the process of selfregistration and sharing a shelf, I have adjusted
the text "has been sent" to "will be sent shortly". This covers imo
the one-hour frequency.

When writing this patch, I have examined all calls of EnqueueLetter;
I only found these two occurrences to be of interest.

Note: I would recommend to increase this frequency in the documentation,
but consider that for now to be outside the scope of this report.

Test plan:
[1] Self-register a new user with verification by email required. Look at
    the text when you submit your data.
[2] Share a list with someone else. Look at the text when you submit the
    invitation.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as described, small string change.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>