Bug 13120 - Invalid HTML generated by MARC21slim2OPACResults.xsl
This patch converts an id to a class in order to avoid duplicate id
attributes on one page.

To test you must have XSLT enabled with the default setting. Perform a
search in the OPAC and confirm that the display of location looks
correct.

Run the resulting page through a validator using an in-browser tool like
Firefox's Html Validator addon or using the W3C's validator.w3.org. The
results should not complain of duplicate id attributes.

Signed-off-by: Frederic Demians <f.demians@tamil.fr>

OK for me. Obvious mistake: I can confirm FF HTML validator reports it, and
stop reporting it after applying this patch.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
I can confirm this fix is correct.
Also, I found that jQuery doesn't like repeated ids in some cases
so this will make customizing easier.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>